# wsfl-frontend/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "WSFL (Winter Series Fantasy League) API",
        "description": "API for managing fantasy running league competitions",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com",
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://choosealicense.com/licenses/mit/"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Development server"
        }
    ],
...
    "paths": {
...
        "/api/races/{id}": {
            "get": {
                "tags": [
                    "Race"
                ],
                "summary": "Get race by ID",
                "description": "Retrieves detailed information about a specific race",
                "operationId": "getRaceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved race",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RaceDetailsDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Race not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Race not found with id: 1"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Race"
                ],
                "summary": "Update race",
                "description": "Updates an existing race's information",
                "operationId": "updateRace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RaceUpdateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Race updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RaceDetailsDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Invalid input data"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Race or League not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Race not found with id: 1"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Race"
                ],
                "summary": "Delete race",
                "description": "Deletes a race from the system",
                "operationId": "deleteRace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Race deleted successfully"
                    },
                    "404": {
                        "description": "Race not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Race not found with id: 1"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/races": {
            "get": {
                "tags": [
                    "Race"
                ],
                "summary": "Get all races",
                "description": "Retrieves a paginated list of all races in the system",
                "operationId": "getAllRaces",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved races",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageRaceSummaryDto"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Race"
                ],
                "summary": "Create new race",
                "description": "Creates a new race in the system",
                "operationId": "createRace",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RaceCreateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Invalid input data"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "League not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "League not found with id: 1"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Race created successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RaceDetailsDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/races/{id}/results": {
            "get": {
                "tags": [
                    "Race"
                ],
                "summary": "Get race results",
                "description": "Retrieves all results for a specific race",
                "operationId": "getRaceResults",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Race not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Race not found with id: 1"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully retrieved results",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageRaceResultSummaryDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/races/{id}/matchups": {
            "get": {
                "tags": [
                    "Race"
                ],
                "summary": "Get matchups by race",
                "description": "Retrieves a paginated list of matchups for a specific race",
                "operationId": "getMatchupsByRace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the race",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "pageable",
                        "in": "query",
                        "description": "Page number (0-based)",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    },
                    {
                        "name": "teamId",
                        "in": "query",
                        "description": "Filter by team ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Race not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Race not found with id: 123"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successfully retrieved matchups",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageMatchupSummaryDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RaceUpdateDto": {
                "required": [
                    "date",
                    "isPlayoff",
                    "leagueId",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the race",
                        "example": "Spring Championship"
                    },
                    "date": {
                        "type": "string",
                        "description": "Date of the race",
                        "format": "date",
                        "example": "2024-06-15"
                    },
                    "isPlayoff": {
                        "type": "boolean",
                        "description": "Whether this is a playoff race",
                        "example": false
                    },
                    "leagueId": {
                        "type": "integer",
                        "description": "ID of the league this race belongs to",
                        "format": "int64",
                        "example": 1
                    }
                },
                "description": "Data transfer object for updating a race"
            },
            "RaceDetailsDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Race's unique identifier",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the race",
                        "example": "Spring Championship"
                    },
                    "date": {
                        "type": "string",
                        "description": "Date of the race",
                        "format": "date",
                        "example": "2024-06-15"
                    },
                    "isPlayoff": {
                        "type": "boolean",
                        "description": "Whether this is a playoff race",
                        "example": false
                    },
                    "league": {
                        "$ref": "#/components/schemas/LeagueSummaryDto"
                    },
                    "results": {
                        "type": "array",
                        "description": "List of race results",
                        "items": {
                            "$ref": "#/components/schemas/RaceResultSummaryDto"
                        }
                    }
                },
                "description": "Detailed representation of a race"
            },
...
            "RaceCreateDto": {
                "required": [
                    "date",
                    "leagueId",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the race",
                        "example": "Spring Championship"
                    },
                    "date": {
                        "type": "string",
                        "description": "Date of the race",
                        "format": "date",
                        "example": "2024-06-15"
                    },
                    "isPlayoff": {
                        "type": "boolean",
                        "description": "Whether this is a playoff race",
                        "example": false
                    },
                    "leagueId": {
                        "type": "integer",
                        "description": "ID of the league this race belongs to",
                        "format": "int64",
                        "example": 1
                    }
                },
                "description": "Data transfer object for creating a new race"
            },
...
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "SortObject": {
                "type": "object",
                "properties": {
                    "empty": {
                        "type": "boolean"
                    },
                    "sorted": {
                        "type": "boolean"
                    },
                    "unsorted": {
                        "type": "boolean"
                    }
                }
            },
...
            "PageRaceSummaryDto": {
                "type": "object",
                "properties": {
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RaceSummaryDto"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "PageRaceResultSummaryDto": {
                "type": "object",
                "properties": {
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RaceResultSummaryDto"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}